requirements.txt
azure-batch==6.0.0
azure-storage-blob==1.3.1
coloredlogs